+++ cli.js
@@ -1,8 +1,10 @@
 #!/usr/bin/env node
 
-const config = {};
+const config = {
+    views: {}
+};
 
-for (let i = 2; i < process.argv.length - 1; i++) {
+for (let i = 2; i < process.argv.length; i++) {
     switch (process.argv[i]) {
         case "-h":
         case "--handler":
@@ -21,6 +23,22 @@ for (let i = 2; i < process.argv.length - 1; i++) {
             if (!folder.endsWith("/")) folder += "/";
             config.key = folder + "ssl.key";
             config.cert = folder + "ssl.cert";
+            break;
+        case "--view-engine":
+            config.views.engine = process.argv[++i];
+            break;
+        case "--views-directory":
+            config.views.directory = process.argv[++i];
+            break;
+        case "--trust-proxy":
+            config.trustProxy = true;
+            break;
+        case "--bcrypt-bench":
+            config.bcryptBench = true;
+            break;
+        case "--dotenv":
+            config.dotenv = true;
+            break;
     }
 }